requirements.txt
pyodbc
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
